requirements.txt
altair==4.2.2
annotated-types==0.6.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.3
branca==0.6.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
dacite==1.8.1
decorator==5.1.1
deprecation==2.1.0
duckdb==0.8.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.0.0
Faker==19.11.0
favicon==0.7.0
folium==0.14.0
fonttools==4.43.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.0
gw_dsl_parser==0.1.6a6
htbuilder==0.6.1
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.8.0
ipython==8.16.1
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.1.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
llvmlite==0.41.1
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
more-itertools==10.1.0
multimethod==1.10
mysql-connector-python==8.2.0
networkx==3.2
numba==0.58.1
numerize==0.12
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==10.1.0
Plost==0.2.5
plotly==5.14.1
plotly-express==0.4.1
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.6
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pygwalker==0.3.9
pymdown-extensions==10.3.1
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
scikit-learn==1.2.2
scipy==1.11.3
seaborn==0.12.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.22
sqlglot==18.16.1
st-annotated-text==4.0.1
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.28.0
streamlit-aggrid==0.3.4.post3
streamlit-apexjs==0.0.3
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-elements==0.1.0
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-folium==0.15.0
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-modal==0.1.0
streamlit-option-menu==0.3.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
typeguard==4.1.5
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
vincent==0.4.4
visions==0.7.5
wasmtime==12.0.0
watchdog==3.0.0
wcwidth==0.2.8
widgetsnbextension==4.0.9
wordcloud==1.9.2
ydata-profiling @ https://github.com/pandas-profiling/pandas-profiling/archive/master.zip#sha256=46d263b6caff9a80d5fa2d4aed18d2c7566b1bd9714e890130d8b49834de184c
zipp==3.17.0
